all imported dependencies moved to deps.ts file
