--- Test/deps.ts
@@ -1,16 +1,0 @@
-export {
-  assert, 
-  assertEquals,
-  assertExists,
-  assertNotEquals,
-  assertStrictEquals, assertThrows, 
-} from "https://deno.land/std@0.150.0/testing/asserts.ts";
-
-export {
-  afterAll,
-  afterEach,
-  beforeAll,
-  beforeEach,
-  describe,
-  it
-} from "https://deno.land/std@0.150.0/testing/bdd.ts";
