+++ deps.ts
@@ -4,4 +4,31 @@ export { config, parse } from "https://deno.land/std@0.148.0/dotenv/mod.ts"; // 
 
 export { Pool, PoolClient } from "https://deno.land/x/postgres@v0.17.0/mod.ts"; // Db.ts
 
-export { ensureDir } from "https://deno.land/std@0.165.0/fs/ensure_dir.ts"; // init.ts
+export { ensureDir } from "https://deno.land/std@0.165.0/fs/ensure_dir.ts"; // init.ts, checkDbPull.ts, checkDbSync.ts
+
+export {resolve} from "https://deno.land/std@0.170.0/path/mod.ts"; //checkDbSync.ts, dbPull.ts, modelParser.ts, seed.ts, checkDbPull.ts, 
+
+export {join} from "https://deno.land/std@0.141.0/path/win32.ts"; // introspect.ts
+
+export {readLines} from "https://deno.land/std@0.141.0/io/buffer.ts"; // myLog.ts, restore.ts
+
+export {
+  assert, 
+  assertEquals,
+  assertExists,
+  assertNotEquals,
+  assertStrictEquals, assertThrows, 
+} from "https://deno.land/std@0.150.0/testing/asserts.ts";
+
+export {
+  afterAll,
+  afterEach,
+  beforeAll,
+  beforeEach,
+  describe,
+  it
+} from "https://deno.land/std@0.150.0/testing/bdd.ts";
+
+// unknown how to export : import "https://deno.land/x/dotenv@v3.2.0/load.ts"; which is imported on files: ./mod.ts, ./src/Db.ts, ./Test/dbPing.ts, ./Test/association.ts
+
+// unused import { resolve } from "https://deno.land/std@0.155.0/path/win32.ts" in files: ./mod.ts, ./src/functions/restore.ts
